+++ apps/api/src/games/games.module.ts
@@ -8,12 +8,13 @@ import { AuditLogModule } from '../audit-log/audit-log.module';
 import { FeedModule } from '../feed/feed.module';
 import { GamesController } from './games.controller';
 import { GamesService } from './games.service';
+import { GamePublicationService } from './game-publication.service';
 import { CountersService } from '../common/counters.service';
 
 @Module({
   imports: [PrismaModule, AuthModule, StudiosModule, AuditLogModule, FeedModule, AnalyticsModule],
   controllers: [GamesController],
-  providers: [GamesService, CountersService],
+  providers: [GamesService, GamePublicationService, CountersService],
   exports: [GamesService],
 })
 export class GamesModule {}